+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+
+## [1.7.0] - 2024-04-10
+### New
+- Added `getExchangeOffers` endpoint into DataUtils
+### Breaking changes
+- The type name `UserOpsTransaction` has been changed to `UserOpTransaction`
+
 ## [1.6.5] - 2024-04-03
 ### New
 - Added Rootstock testnet and mainnet network support